Add capability for cron container to delete all locks
Add the capability for the cron container to delete all
locks in the table cm_locks. This will allow a sidecar
container running next to the cron container with the command
/usr/local/bin/delete-all-locks.sh to receive a SIGTERM
and delete all locks before the sidecar shuts down. Note that
Kubernetes will send the SIGTERM to the main container and wait
for it to exit before sending the SIGTERM to the sidecar, so the
sidecar should be able to clean up all locks.